Speed up the generate_carrier_list function: --- **Changes:**
*generate_carrier_list:* changed this function so that instead of
iterating over each column it transposes the matrix and then applies a
function called _identify_carriers_indx to find the list of indices for
affected individuals for each phecode.

*_identify_carriers_indx:* function that will add the indices list of
each affect individuals as values to a dictionary where the keys are the
phecodes. This dictionary is created using the apply method in the
generate_carrier_list function