Creating a numerical id for grids and removing the PheCode columns: ---
1. **Creating integer ids for the grid ids:**

To save memory and potentially speed up operations I am switching over
the grid ids to just use integers that correspond to the grid's position
in the dataframe. These ids start with 0 for the first grid and then
increase by 1. There are ids for both haplotypes so grid1.1 will have
one value while grid1.2 will have a separate value. This is part of a
refactoring design that needs to be tested because it may not actually
speed up any computations.

2. **Removing the phenotype columns:**

I removed the phecode columns of 0's/1's that gets added to the ibd_df
attribute of the cluster object because it takes a lot of space and
makes the dataframe have a much higher memory consumption which affects
how the program can be run in parallel.

* *Things that need to still be done for this part.*

The phenotype information is still useful (and used) in the binomial
test statistic. This means that there needs to be some data structure
for the program to keep teh information. The program only needs to keep
track of who carries what phecode and it doesn't need to know who
isn't affect so this should reduce redundancy and therefore memory
consumption.
2.