Removing the Hapibd and iLASH fileIndex models ___
**Added a ibd_file argument flag to the typer app**

This allows users to specify which file they need and it removes a lot of
redundant and unnecessary code. This change also moves the
responsibility of making sure the gene info chromosome and the ibd file
chromosome number match from the program to the user. Since it was not
being parallelized then it made more since to just have the user give
the exact file.

**Remove the HapibdInfo and the IlashInfo protocals**

Removed these two protocols. Instead moved the index into the
set_cM_indx function in the FileInfo class. This function uses a match
case to match either hapibd or ilash or raise a NotImplemented Error.

**Also changed the type for the ibd_program flag for the
IBDCluster.py/main function:**

Now the flag uses an enum with types for hapibd and ilash. It raises an
error if the value is not in the enum and it is case sensitive.