Optimization of the pvalue.py: ---
1. Use Set intersection instead of list comprehension to determine who
is affected with a phenotype.

Changed line 115 which was a list comprehension to a set intersection to
find who in the network.iids set is also in the carriers set for the
phenotype (This code is now found at line 113). This gave a speed up
from ~2500s to 8s as measured using cProfile and Snakeviz.

2. Other changes.

Also removed unecessary log statements and commented out lines that were
not still used (line 116,118) that could be added back later but they
don't need to be there now