Added a connection threshold: ---
Added a flag for a connection threshold. This was done to break
connections in large networks that may be artificially inflating the
size of networks. The flag defaults to 0. Updated the functions that
require this and then adjusted the program so that it filters the
self.ibd_df for only individuals that >= the threshold. This change
improves the runtime because it makes the search space smaller