Switched the theme and added information about installation: --- 1. Switched the theme to Just the Docs:
This theme is used by SAIGE and is pretty nice and came with a search
box and a nav bar so it simplified my life.

2. Added a link to the github page:

This link is found in the upper right hand corner. I don't love the font
but it works for the time being.

3. Started creating  documentation section:

Creating files for the inputs and the how to Install the program. Right
now the two installation methods are github and Pip. The github is a
work in progress but there is information there. The pip installation is
not actually implemented yet but it is there because it is part of the
roadmap.