Created documentation about the API: --- 1. Added info about the API:
Listed all of the different models in the API. These models are the
DataHolder model, the Network model, and the Pairs model. Also described
all of the attributes and the methods of each attributes. Also described
how the api provides the output path that the plugin can write to.

2. Changing the design_your_own.md name to design_your_own_plugin.md:

Renamed the file to make it more clear what the file is about.