Added Info about stock plugins and the program inputs: --- 1. Info about stock plugins:
Added info to the Stock Plugins section that goes over how to find the
default plugins, and then a general overview of the plugin structure,
and then the description of the three stock plugins: pvalues.py,
network_writers.py, and allpairs_writer.py.

2. Info about program inputs:

Added info about the ibd file, the config.json file, the .env file, and
the Gene Information text file. There are tables that show the expected
columns when necessary and examples of what the file should look like.