Added a sliding window: ---- 1. Added the ability for the user to implement a sliding window approach to determine loci of interest:
The user can now pass the flag --sliding-window to the program. This
value defaults to false. If true then the program expects the gene info
file to just have 1 line and the program will create a range from the
loci start to loci end with a step size of 1000. This range is inclusive
of the end. The program then creates a Genes namedtuple that it uses in
the genes_generator.

2. Updated the cluster.load_gene_info function:

This function now accepts an argument called sliding_window which will
be either true or false. The value of this argument is checked in a match
statement. If True, then the program will read the only line in the
genes file in to memory and then extract the full range of the loci. It
will then create 1 MB steps to the inclusive end of the range. Note: the
final step may not be 1 MB, but is not expected to be > 1 MB. It will
then yield a Genes namedTuple where the name is now the first column in
the genes info file with the range for each window appended to it. If
sliding_window == False then the program behaves as before.