Removed the need for the gene info file: --- 1. Removed the gene info file input:
Replaced this flag with 2 flags: gene_position and gene_name. These can
be provided by the user. The motivation is that the program was not
designed in a way to run multiple gene targets from the gene info file
so this file was restricted to one line anyways and was unnecessarily.
This input is recorded in the log file so the user can still tell what
they ran. The --gene-position flag expects the user to provided a string
of the format "chromosome:start_position-end-position". The --gene-name
flag allows the user to give a name to the gene. This flag defaults to
test if the user doesn't provided anything.

2. Updated the load_gene_info function in the cluster.main file:

This function now returns a list of Genes instead of a generator. The
function uses a regular expression to pull out the chromosome number and
start/end positions from the provide chromo_pos_str. Still retains the
ability to create a sliding window every 1MB.

3. Add a callback to check the format of the --gene-position string:

Added a function called check_gene_pos_str in callbacks.check_inputs.py.
This function makes sure that the string is formatted as
"chromosome:start_position-end_position". This format is checked using
the re.split to make sure that the resulting split list is of length 3.
If there is a format issue than a ValueError is raised with a message

Things to do:
---
These changes broke the unit test for the sliding window function. So
this needs to be fixed