refactor(drive/callbacks.py): created callback function to make sure that the input ibd file exist
Created a function called check_input_exists that will either return the input Path object or it will raise a FileNotFoundError if the file doesn't exists. It is passed as a callback to the typer option input and takes whatever the user passes for that option. Need to also check if this input file is gzipped at some point.