refactor(drive/DRIVE.py): refactor how the target string was split and how to indices are created
Created a function called split_target_string that takes the target argument passed by the user (line 169). This function uses a regex expression to split the string. If it doesn't split the string into a list of 3 strings then it raises a value error indicating that the target was not the proper format. Otherwise it returns a Genes object named tuple that has the attributes chr, start, and end. Also started to convert lines 129-166 from a lot of branch to different classes that have similar methods. This approach may be changed in the future (will talk to alex about this). Also dropped support for rapid at the moment. This was not used so its not really a breaking change. Once I decide if the class based approach is right then I will add it back.