test(pyproject.toml-and-tests): started building the integration test
Installed pytest as a development dependency (v7.2.2). Then added a function to the test_integration.py file called test_drive_full_run() that will run the program and check if the exit code is 0. This was done to ensure reproducibility for the KCNE1 paper. I am waiting for hung-hsin to give me the exact command he used so that it can reproduce.