refactor(file-indices): Switched to a strategy pattern
replaced lines 89-126 with a strategy pattern design. Created a file called generate_indices.py that has class for each of the ibd programs. These class all have the getHAPID method but implement it different. Also created an Interface defining that the object needs to implement the getHAPID method. Created a function called create_indices that decides which object to create and then returns the proper object or raises a valueError if the user passed an unsupport format type. This funciton is called in the DRIVE.py script