refactor(drive/generate_indices.py): remove branching in create_indices
Removed the if/else statements in the generate_indices.create_indices function. Replaced this with a dictionary and then the method just searches the dictionary keys. To ensure that there would not be a key error I also used a Enum call FormatTypes as the type for the format option. This enum ensures that users can only pass hapibd, ilash, germline, or rapid.