refactor(filtering-ibd-file): Restructured hung-hsin's filtering into a module
Created a filter module that has a Filter class. This class replaces all the filtering hung-hsin did in if statements with vectorized code in pandas dataframes. The class stores the ibdpd and ibdvs as attributes called ibd_pd and ibd_vs. If the user provides an IBD file that doesn't exist than it will raise a FileNotFoundError.