refactor: Starting refactoring the clustering and reorganizing the models
Created a module called cluster with a cluster.py file. This file will have the main clustering function and a class called networks that will be respon
ible for performing the clustering. Also restructured the protocols and named tuple so that they are in the models cluster now and the classes can just beimported.