refactor(clustering): Broke the Networks class into two classes
Broke the networks class into a Networks class and the ClusterHandler class. This change takes some responsibility away from the Networks class. Now the ClusterHandler is responsible for the random walk and getting information about the clusters. The ClusterHandler class uses the graph in the Networks class to perform the walk. It will also add each cluster to the Networks.clusters attribute dictionary