refactor(PhenotypeFileParser): added functionality to the parser so it can determine cases/controls/exclusions
added a method called parse_cases_and_controls that reads through the input file and separates individuals into case/controls/exclusion lists. This method returns those list
so that the user can later use them. Created unit test that makes sure that the parser raises the proper exceptions and that it identifies the right number of individuals based on the test data