refactor: Added attribute to keep a list of all individuals in cohort
Added an attribute called individuals that keep track of which grids should be in the analysis. This attribute will be used to filter the ibd file to only those individuals we want. Not filter could cause a bug by introduce extra individuals into the network size that could affect the stats. Also fixed a bug where individuals that had an unrecognized status were not being added to the proper excluded list.