refactor: refactor filter to remove unnecessary function and to determine all individuals in cohort
Added an attribute called individuals that is a list of all individuals in the cohort. This list is used to filter the ibd file for only individuals in our cohort. Also removed function to determine how many chunks the dataframe is split into. This function was used to create a progress bar but that is no longer necessary.