refactor: make sure that the ibd file matches the chromosome of the chromosome target
If the user passed a ibd file that was different than the chromosome from the target string then an error would be raised later in the program when a column in an empty dataframe was trying to be accessed by the idnum attribute. Added an error check that would catch if the chromosome number from target_gene was not in the dataframe. This situation now raises a ValueError that will tell the user they should check the provided file