fix(network_writer): correct how the output file name is made when it contains a period
There was a bug that when the output name had a "." already in it then the program would consider that a suffix and then replace it. This was causing bugs for Phecodes which has subtypes such as 450 450.2, etc... Switched from using the Path().with_suffix() to using / and (x.name + "suffix.txt") syntax which should fix it.