fix(cluster.py): fixed a bug that caused the IDs to mismatch
There was a bug where the member_list and the vertex_ids were being flipped incorrectly. This bug was causing the incorrect ids to be mapped back on to grids when the reclustering weas performed. Now that is fixed and results have been compared to the KCNE1 results from DRIVE 1.0.0 to maintain backward compatibility