feat(case_file_parser.py): added ability to specify one phenotype column in a matrix
Added a new property to the PhenotypeFileParser called specific_phenotype. This property allows the user to provided a specific phenotype name to the drive program using the argument '--phenotype-name'. This change allows the user to specify a specific phenotype column from a file without having to recreate a bunch of phenotype files.