fix(case_file_parser.py): Fixed a bug where the case file parser was not identifying the appropriate index if the user only wished to find a specific haplotype
The parser would attempt to get the appropriate index of where the phenotype value was in the list. By default it starts at zero but if the user only want to specify a specific column then a bug appeared where if the user specified a column such as 10, the parser would attempt to first get the value at index 0 which would return a none value. This none value would then cause the latter code to fail.