fix(ClusterHandler.redo_clustering): bug caused when no individuals shared pairwise IBD segments and graph was improperly formed
There was an edge case that when the redopd was formed it would be empty. This emptiness would be caused when none of the individuals in network.haplotypes shared pairwise segments. This empty dataframe would then be used in the generate_graph function and instead of failing the ig.Graph.DataFrame constructor would return an object with an empty list for the edgelist (.es) attribute. The Graphbase.community_walktrap function would later on try to use this attribute and since it was empty the code would fail. Now DRIVE checks to see if the redopd or redo_vs dataframes are empty. If they are empty DRIVE will continue onto the next network, except in debug mode where DRIVE will give a logging message explaining that a graph could not be constructed during the reclustering of that network.